Extract ReactComponentTreeDevtool
